pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "version-watch"
version = "0.1.0"
description = "API versioning manager: detect breaking changes, track deprecations, generate sunset plans"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.9"
authors = [{name = "version-watch contributors"}]
keywords = ["api", "versioning", "openapi", "breaking-changes", "deprecation"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Topic :: Software Development :: Libraries",
]
dependencies = [
    "click>=8.0",
    "pyyaml>=6.0",
    "rich>=13.0",
]

[project.scripts]
version-watch = "version_watch.cli:cli"

[tool.setuptools.packages.find]
include = ["version_watch*"]
